add restart, reset button and logic.
