+++ index.html
@@ -27,7 +27,9 @@
                 <tbody id="tbody">
                 </tbody>
             </table>
-            <div class='d-flex w-100 justify-content-center'><button class="btn btn-info" id='reset'>Reset Game</button></div>
+            <div class='d-flex w-100 justify-content-center'>
+              <button class='btn btn-primary mr-2' id='restart'>Restart Game</button>
+              <button class="btn btn-info" id='reset'>Reset Game</button></div>
         </div>
         <script type="module" src="assets/index.js"></script>
     </body>
